tratamento de excessos no delete

+++ src/main/java/com/zaninweb/course/services/UserService.java
@@ -1,14 +1,16 @@
 package com.zaninweb.course.services;
 
-
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.zaninweb.course.entities.User;
 import com.zaninweb.course.repositories.UserRepository;
+import com.zaninweb.course.services.exceptions.DatabaseException;
 import com.zaninweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -16,26 +18,33 @@ public class UserService {
 
 	@Autowired
 	private UserRepository repository;
-	
+
 	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
-	public User findById(Long id) {
-	    Optional<User> obj = repository.findById(id);
-	    return obj.orElseThrow(() -> new ResourceNotFoundException(id));
-	}
 
+	public User findById(Long id) {
+		Optional<User> obj = repository.findById(id);
+		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
+	}
 
 	public User insert(User obj) {
-		
+
 		return repository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
 		User entity = repository.getReferenceById(id);
 		updateData(entity, obj);
@@ -46,8 +55,7 @@ public class UserService {
 		entity.setName(obj.getName());
 		entity.setPhone(obj.getPhone());
 		entity.setEmail(obj.getEmail());
-	
+
 	}
-	
-	
+
 }
